Correcting rough fit: error estimation and binning
Corrected some problems with the average neutrino correction. The
binning is now finer in the low-E region to account for higher
statistics. In the error estimation the estimator was changed from "std
dev of mean" to "std dev of distribution" ( = Need to account for shape
uncertainty and not for low statistics in my MC).